Move execution/history stage inserts to provider
