+++ crates/executor/src/lib.rs
@@ -10,8 +10,8 @@
 pub mod config;
 pub mod eth_dao_fork;
 
-/// Execution result types
-pub mod execution_result;
+/// Execution result types.
+pub use reth_provider::execution_result;
 /// Executor
 pub mod executor;
 /// Wrapper around revm database and types
